+++ src/main/java/org/improvejava/kurento_chat/dto/CreateRoomDTO.java
@@ -1,20 +1,21 @@
-package org.improvejava.kurento_chat;
+package org.improvejava.kurento_chat.dto;
 
 import lombok.Getter;
+import org.improvejava.kurento_chat.utils.RoomIdGenerator;
 
 import java.util.UUID;
 
 @Getter
-public class JoinRoomDTO {
+public class CreateRoomDTO {
     private final String userId;
 
     private final String userName;
 
     private final String roomId;
 
-    JoinRoomDTO(String userName, String roomId) {
+    CreateRoomDTO(String userName) {
         this.userId = UUID.randomUUID().toString();
         this.userName = userName;
-        this.roomId = roomId;
+        this.roomId = RoomIdGenerator.generateRoomId();
     }
 }
