+++ src/main/java/org/improvejava/kurento_chat/dto/JoinRoomDTO.java
@@ -1,20 +1,20 @@
-package org.improvejava.kurento_chat;
+package org.improvejava.kurento_chat.dto;
 
 import lombok.Getter;
 
 import java.util.UUID;
 
 @Getter
-public class CreateRoomDTO {
+public class JoinRoomDTO {
     private final String userId;
 
     private final String userName;
 
     private final String roomId;
 
-    CreateRoomDTO(String userName) {
+    JoinRoomDTO(String userName, String roomId) {
         this.userId = UUID.randomUUID().toString();
         this.userName = userName;
-        this.roomId = RoomIdGenerator.generateRoomId();
+        this.roomId = roomId;
     }
 }
